Simplified bayesian exploration and modified output to dict
